Extract BehaviorToolConfig into tool_surface/behavior_config.rs
